converted to unsigned longs

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -9,9 +9,9 @@
 int main(void)
 {
 	int i;
-	int fib_one = 1; /* fib number 1 */
-	int fib_two = 2; /* fib number 2 */
-	int next = fib_one + fib_two; /* result of 1+2 */
+	unsigned long int fib_one = 1; /* fib number 1 */
+	unsigned long int fib_two = 2; /* fib number 2 */
+	unsigned long int next = fib_one + fib_two; /* result of 1+2 */
 
 	printf("%d, %d, ", fib_one, fib_two);
 	for (i = 3; i < 50; ++i)
